+++ app/schemas/requests.py
@@ -1,0 +1,17 @@
+from pydantic import BaseModel, Field
+from typing import Optional
+from .common import CaregiverType
+
+class OnboardingRequest(BaseModel):
+    name: str = Field(..., min_length=1, max_length=50, description="사용자 이름")
+    age: int = Field(..., ge=1, le=120, description="사용자 나이")
+    caregiver_type: CaregiverType = Field(..., description="부양자 유형")
+    caregiver_age: int = Field(..., ge=1, le=120, description="부양자 나이")
+
+class MessageRequest(BaseModel):
+    user_id: str
+    message: str
+
+class WebSocketMessage(BaseModel):
+    message: str
+    user_id: Optional[str] = None
